requirements.txt
Package           Version
----------------- -----------
alembic           1.12.0
bcrypt            4.0.1
beautifulsoup4    4.12.2
blinker           1.6.3
cffi              1.16.0
click             8.1.7
colorama          0.4.6
cryptography      41.0.4
dnspython         2.4.2
email-validator   2.0.0.post2
Flask             2.3.0
Flask-Bcrypt      1.0.1
Flask-Login       0.6.2
Flask-Mail        0.9.1
Flask-Migrate     4.0.5
Flask-SQLAlchemy  3.1.1
Flask-WTF         1.2.1
greenlet          3.0.0
idna              3.4
itsdangerous      2.1.2
Jinja2            3.1.2
Mako              1.2.4
MarkupSafe        2.1.3
Pillow            10.1.0
pip               23.2.1
pycparser         2.21
PyJWT             2.8.0
PyMySQL           1.1.0
setuptools        68.2.2
soupsieve         2.5
SQLAlchemy        2.0.22
typing_extensions 4.8.0
Werkzeug          2.3.0
wheel             0.41.2
WTForms           3.1.0
